% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dqs_scores.R
\name{extract_dqs_scores}
\alias{extract_dqs_scores}
\title{Extract Diet Quality Score (DQS) Data from an eNutri Export}
\usage{
extract_dqs_scores(enutri_source_file)
}
\arguments{
\item{enuti_source_file}{an absolute file path of a \code{.xlsx} workbook}
}
\value{
a \code{tbl_df} of eNutri diet quality score data
}
\description{
Extract Diet Quality Score (DQS) Data from an eNutri Export
}
