% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_food_gday.R
\name{extract_food_gday}
\alias{extract_food_gday}
\title{Extract Grams per Day Frequency Data from an eNutri Export}
\usage{
extract_food_gday(enutri_source_file)
}
\arguments{
\item{enuti_source_file}{an absolute file path of a \code{.xlsx} workbook}
}
\value{
a \code{tbl_df} of all eNutri grams per day data
}
\description{
Extract, clean and compile all grams per day related data from a eNutri workbook export. This functions combines g/day data that is exported as three different sheets
within a workbook and collates data into a single \code{tbl_df} with the correct variable types
}
\examples{
\dontrun{
enutri_export <- 'eNutri-export-ffq-01012026.xlsx'

enutri_gday_data <- extract_food_gday(enutri_export)

}
}
