% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_nutrients.R
\name{extract_nutrients}
\alias{extract_nutrients}
\title{Extract Nutrient Data from an eNutri Export}
\usage{
extract_nutrients(enutri_source_file, output)
}
\arguments{
\item{output}{a character of one of the following three options;
\itemize{
\item \code{abs}; returns all available nutrient data as absolute values (ie, \code{g}, \code{mg}, \code{ug} etc...)
\item \code{te}; returns all nutrient data that is available as a percentage of total energy intake
\item \code{extra}; returns all additional data that is not a specific nutrient, ie \verb{Greenhouse gas emissions (kg CO2 eqv)} and \verb{Acidification (g SO2 emissions)}
}}

\item{enuti_source_file}{an absolute file path of a \code{.xlsx} workbook}
}
\value{
a \code{tbl_df} of eNutri nutrient data
}
\description{
Extract Nutrient Data from an eNutri Export
}
